requirements.txt
pytest==8.4.1
Appium-Python-Client==5.1.1
ftd2xx==1.3.8
pyftdi==0.56.0